Fix review issues and also checking numeric value for bonus price salesFrequency and sort
