+++ src/FINDOLOGIC/Export/Data/Price.php
@@ -2,14 +2,19 @@
 
 namespace FINDOLOGIC\Export\Data;
 
+use FINDOLOGIC\Export\Helpers\UsergroupAwareNumericValue;
 use FINDOLOGIC\Export\Helpers\UsergroupAwareSimpleValue;
 
 class Price extends UsergroupAwareSimpleValue
 {
-    const PRICE = true;
 
     public function __construct()
     {
-        parent::__construct('prices', 'price', self::PRICE);
+        parent::__construct('prices', 'price');
+    }
+
+    static function validate($value)
+    {
+        return UsergroupAwareNumericValue::validate($value);
     }
 }
