+++ src/FINDOLOGIC/Export/Data/SalesFrequency.php
@@ -3,11 +3,17 @@
 namespace FINDOLOGIC\Export\Data;
 
 use FINDOLOGIC\Export\Helpers\UsergroupAwareSimpleValue;
+use FINDOLOGIC\Export\Helpers\UsergroupAwareNumericValue;
 
 class SalesFrequency extends UsergroupAwareSimpleValue
 {
     public function __construct()
     {
         parent::__construct('salesFrequencies', 'salesFrequency');
+    }
+
+    static function validate($value)
+    {
+        return UsergroupAwareNumericValue::validate($value);
     }
 }
