+++ src/FINDOLOGIC/Export/Helpers/UsergroupAwareSimpleValue.php
@@ -2,6 +2,14 @@
 
 namespace FINDOLOGIC\Export\Helpers;
 
+class EmptyValueNotAllowedException extends \RuntimeException
+{
+    public function __construct($message = 'Empty values are not allowed!')
+    {
+        parent::__construct($message);
+    }
+}
+
 /**
  * Class UsergroupAwareSimpleValue
  * @package FINDOLOGIC\Export\Helpers
@@ -13,13 +21,11 @@ abstract class UsergroupAwareSimpleValue implements Serializable
     private $collectionName;
     private $itemName;
     private $values = array();
-    private $price;
 
-    public function __construct($collectionName, $itemName, $price = false)
+    public function __construct($collectionName, $itemName)
     {
         $this->collectionName = $collectionName;
         $this->itemName = $itemName;
-        $this->price = $price;
     }
 
     public function getValues()
@@ -32,10 +38,30 @@ abstract class UsergroupAwareSimpleValue implements Serializable
      */
     public function setValue($value, $usergroup = '')
     {
-        $this->values[$usergroup] = DataHelper::checkForEmptyValue($value);
-        if ($this->price === true) {
-            DataHelper::checkForValidPrice($value);
+        $this->values[$usergroup] = $this->validate($value);
+    }
+
+    /**
+     * Validates given value.
+     * Basic implementation is validating against an empty string,
+     * but is overridden when checking values more specific.
+     *
+     * When valid returns given value.
+     * When not valid an exception is thrown.
+     *
+     * @param $value string|int Validated value.
+     * @return string string|int
+     * @throws EmptyValueNotAllowedException
+     */
+    static function validate($value)
+    {
+        $value = trim($value);
+
+        if ($value === '') {
+            throw new EmptyValueNotAllowedException();
         }
+
+        return $value;
     }
 
     /**
